Example app: * Factory_MVC::getViewInstance is strictly used instead of direct view instance creation. * ACL action constant names unification * Admin menu UI updated
Jet Studio:
* ACL action constant names unification - module wizard templates
* Better module wizard templates directory organization

Jet library (core):
* Translator::setCurrentDictionaryTemporary can return value returned by callback
* Form_Field_Date: bug fix (wrong error message code)
* Form_Renderer_Single->setCustomRenderer() added
* Form_Renderer_Single->getCustomRenderer( ?Closure $custom_renderer ):void
* Form_Renderer_Single->renderByView() : string added
* Form_Field_Select and Form_Field_MultiSelect: bug fix - readonly works now
* Form_Field_Int: bug fix - incorrect retyping

+++ library/Jet/Form/Field/Int.php
@@ -36,7 +36,7 @@ class Form_Field_Int extends Form_Field_Input implements Form_Field_Part_NumberR
 		parent::catchInput( $data );
 		
 		if($this->_value!=='') {
-			$this->_value = (float)$this->_value_raw;
+			$this->_value = (int)$this->_value_raw;
 		}
 	}
 }
